+++ test/manual/multipage-watch-partials/src/scripts/hrm.js
@@ -1,1 +1,0 @@
-// to enable HMR add the `<script src="path/to/hrm.js"></script>` to head